requirements_amazon.txt
amazon-braket-default-simulator==1.26.4
amazon-braket-schemas==1.23.1
amazon-braket-sdk==1.90.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.2
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
backports.entry-points-selectable==1.3.0
boltons==25.0.0
boto3==1.37.2
botocore==1.37.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cloudpickle==2.2.1
comm==0.2.2
contourpy==1.3.1
cryptography==45.0.2
cycler==0.12.1
debugpy==1.8.12
decorator==5.2.1
dill==0.4.0
executing==2.2.0
fastjsonschema==2.21.1
fonttools==4.56.0
ibm-cloud-sdk-core==3.23.0
ibm-platform-services==0.66.0
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kaleido==0.2.1
kiwisolver==1.4.8
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy-extensions==1.0.0
narwhals==1.40.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.3
openpulse==1.0.1
openqasm3==1.0.1
opt_einsum==3.4.0
oqpy==0.3.7
packaging==24.2
pandas==2.2.3
parso==0.8.4
pbr==6.1.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.1.0
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydot==4.0.0
Pygments==2.19.1
PyJWT==2.10.1
pylatexenc==2.10
pyparsing==3.2.1
pyspnego==0.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyzmq==26.2.1
qiskit==2.0.1
qiskit-aer==0.17.0
qiskit-ibm-runtime==0.39.0
referencing==0.36.2
requests==2.32.3
requests_ntlm==1.3.0
rpds-py==0.25.0
rustworkx==0.16.0
s3transfer==0.11.3
scipy==1.15.2
seaborn==0.13.2
setuptools==80.8.0
six==1.17.0
stack-data==0.6.3
stevedore==5.4.1
symengine==0.13.0
sympy==1.13.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
